+++ lib/xnix/compatibility/dbus_runtime_client.rb
@@ -129,6 +129,10 @@ module Xnix
         parse_dictionary(call("GetBackendBinding", application_id))
       end
 
+      def backend_lifecycle(application_id)
+        parse_dictionary(call("GetBackendLifecycle", application_id))
+      end
+
       def repair_plan(application_id, issue)
         parse_dictionary(call("GetRepairPlan", application_id, issue))
       end