+++ test/test_runtime_contract.rb
@@ -41,6 +41,7 @@ plasmoid = project_root.join("kde/plasmoids/org.xnix.compatibilitycenter/metadat
   GetCompatibilityArtifactManifest
   GetCompatibilityInstallPlan
   GetBackendBinding
+  GetBackendLifecycle
   GetRepairPlan
   GetTestPlan
   GetTestResult